structure improvement ver1
