+++ app/src/main/java/com/example/aborea/common/SetBackground.kt
@@ -1,0 +1,16 @@
+package com.example.aborea.common
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+@Composable
+fun SetBackground() {
+    Box(modifier = Modifier
+        .fillMaxSize()
+        .background(color = Color(0xFFEEFFDB) )
+    )
+}
