+++ app/src/main/java/com/example/aborea/pages/goal/Goal2.kt
@@ -1,0 +1,12 @@
+package com.example.aborea.pages.goal
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+import com.example.aborea.common.*
+import com.example.aborea.pages.goal.compose.*
+
+@Composable
+fun Goal2(navController: NavController) {
+    SetBackground()
+    NavBar(navController)
+}
