+++ app/src/main/java/com/example/aborea/pages/home/Home1.kt
@@ -1,0 +1,12 @@
+package com.example.aborea.pages.home
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+import com.example.aborea.common.*
+import com.example.aborea.pages.home.compose.*
+
+@Composable
+fun Home1(navController: NavController) {
+    SetBackground()
+    NavBar(navController)
+}
