+++ app/src/main/java/com/example/aborea/pages/statistics/Statistics2.kt
@@ -1,0 +1,12 @@
+package com.example.aborea.pages.statistics
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+import com.example.aborea.common.*
+import com.example.aborea.pages.statistics.compose.*
+
+@Composable
+fun Statistics2(navController: NavController) {
+    SetBackground()
+    NavBar(navController)
+}
