+++ app/src/main/java/com/example/aborea/pages/store/Store2.kt
@@ -1,0 +1,12 @@
+package com.example.aborea.pages.store
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavController
+import com.example.aborea.common.*
+import com.example.aborea.pages.store.compose.*
+
+@Composable
+fun Store2(navController: NavController) {
+    SetBackground()
+    NavBar(navController)
+}
